@@ -13,20 +13,22 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "lint": "eslint --fix src",
     "test:ci": "CI=true react-scripts test --watchAll=false",
+    "lint-staged": "npx lint-staged",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest"
+      "react-app/jest",
+      "prettier"
     ]
   },
   "browserslist": {
@@ -43,7 +45,10 @@
   },
   "packageManager": "yarn@3.6.0",
   "devDependencies": {
+    "eslint-config-prettier": "^8.8.0",
     "jest": "^29.5.0",
-    "jest-watch-typeahead": "^2.2.2"
+    "jest-watch-typeahead": "^2.2.2",
+    "lint-staged": "^13.2.2",
+    "typescript": "5.1.3"
   }
 }
