@@ -1,29 +1,28 @@
-import express, { Express, Request, Response } from 'express';
-const fs = require('fs/promises');
-const path = require('path');
-const FILE_NAME = path.join(__dirname,'../data.json');
-
+import express, { type Express, type Request, type Response } from "express";
+import fs from "fs/promises";
+import path from "path";
+const FILE_NAME = path.join(__dirname, "../data.json");
 
 const app: Express = express();
 
-app.get('/ping', (req: Request, res: Response) => {
-  res.status(200).send('pong');
+app.get("/ping", (_req: Request, res: Response) => {
+  res.status(200).send("pong");
 });
 
-app.get('/ping', (req: Request, res: Response) => {
-  res.status(200).send('pong');
+app.get("/ping", (_req: Request, res: Response) => {
+  res.status(200).send("pong");
 });
 
-app.get('/movies', async(req: Request, res: Response) => {
+// eslint-disable-next-line @typescript-eslint/no-misused-promises
+app.get("/movies", async (_req: Request, res: Response) => {
   try {
-    const file = await fs.readFile(FILE_NAME,{ encoding: 'utf8' });
+    const file = await fs.readFile(FILE_NAME, { encoding: "utf8" });
+
     const data = JSON.parse(file);
     res.status(200).json(data);
-  } catch (e){
+  } catch (e) {
     res.status(500).send(e);
   }
 });
 
-export {app};
-
-
+export { app };
