@@ -5,8 +5,7 @@ import MoviesList from "./components/MoviesList";
 function App() {
   return (
     <div>
-      <header className="App-header">STAR WARS MOVIES</header>
-      <MoviesList></MoviesList>
+      <header className="App-header">STAR WARS MOVIES</header><MoviesList/>
     </div>
   );
 }
