@@ -1,9 +1,10 @@
 import { app } from "./app";
-import dotenv from 'dotenv';
+import dotenv from "dotenv";
 
 dotenv.config();
-const port = process.env.PORT || 3001;
+
+const port = process.env.PORT ?? 3001;
 
 app.listen(port, () => {
-    console.log(`⚡️[server]: Server is running at http://localhost:${port}`);
-}); 
+  console.log(`⚡️[server]: Server is running at http://localhost:${port}`);
+});
